feat(furniture): new feature delete data's and delete images from db & folder

+++ models/furniture/delete.php
@@ -1,9 +1,20 @@
 <?php
 include_once $_SERVER['DOCUMENT_ROOT'].'/core/config.php';
 
-$id = $_GET['id'];
-        
+if (isset($_POST['delete_images_furniture'])) {
+
+    $id = $_POST['delete_id'];
+    $furniture_img = $_POST['delete_images'];
+
     $sql = mysqli_query($conn , "DELETE FROM furnitures WHERE id = '$id'");
 
-    header("location:".url()."resources/furniture");
+    if ($sql) {
+        unlink($_SERVER['DOCUMENT_ROOT']."/public/uploads/".$furniture_img);
+        header("location:".url()."resources/furniture");exit;
+    } else {
+        echo "Gagal";
+    }
+    
+}
+
 ?>
